fix: integer overflow causing segmentation fault On 32-bit architectures this compared UINT64_MAX casted to size_t with UINT64_MAX, which causes an unintended mismatch due to the integer being truncated when calling request_data.

+++ libuuu/buffer.cpp
@@ -1709,7 +1709,7 @@ std::shared_ptr<DataBuffer> FileBuffer::request_data(size_t offset, size_t sz)
 			return p;
 	}
 
-	if (sz == UINT64_MAX)
+	if (sz == SIZE_MAX)
 		sz = size() - offset;
 
 	p->resize(sz);
